pyproject.toml
[project]
name = "presale-auto-python"
version = "0.1.0"
description = "API test framework — Python rewrite (pytest + httpx + pydantic)"
requires-python = ">=3.12"

dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "asyncpg>=0.29",          # аналог pg
    "psycopg[binary]>=3.1",   # синхронный фоллбек если надо
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",     # моки httpx-запросов в тестах
    "pytest-xdist>=3.5",      # параллельный запуск
    "anyio>=4.4",
]

# ---------------------------------------------------------------------------
# dependency-groups (PEP 735, поддерживается uv >= 0.4)
# ---------------------------------------------------------------------------
[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "pytest-xdist>=3.5",
    "anyio>=4.4",
]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode   = "auto"          # все async-тесты без @pytest.mark.asyncio
testpaths      = ["tests"]
python_files   = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Кастомные маркеры — аналог тегов в Playwright
markers = [
    "b2c: тесты для B2C-продукта",
    "b2b: тесты для B2B-продукта",
    "smoke: быстрый smoke-прогон (subset критических)",
    "regression: полный регрессионный прогон",
    "auth: тесты авторизации",
    "orders: тесты заказов",
    "products: тесты продуктов",
    "slow: тесты которые идут дольше 5 сек",
]

# Чтобы pytest не ругался на неизвестные маркеры
addopts = "--strict-markers -ra -q"

# ---------------------------------------------------------------------------
# ruff (опционально, для линтинга)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
